requirements.txt
aniso8601==9.0.1
anyio==3.6.2
beautifulsoup4==4.12.3
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
fastapi==0.93.0
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-Session==0.7.0
h11==0.14.0
httptools==0.6.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
ldap3==2.9.1
MarkupSafe==2.1.5
msgspec==0.18.6
psycopg2-binary==2.9.10
pyasn1==0.5.1
pydantic==1.10.6
pygame==2.6.1
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.25.0
typing_extensions==4.5.0
urllib3==2.2.1
uvicorn==0.21.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
